backend/requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
django-rest-knox==4.2.0
psycopg2-binary==2.9.5
redis==4.4.0
factory_boy==3.2.1
faker==17.0.0